Adding new Browser Labs feature: Slide Tab Transitions
Edge swipe from the sides of the screen to view the previous/next tab in
the same way as the Android homescreen navigates its views. The feature
is enabled by default and can be turned off in Browser settings under
Labs. It is only enabled for the phone UI, it will not run on the tablet
interface. Slide Tab Transitions is incompatible with Quick Actions.

There are still two known issues with the current patch:
- Clipping bug for pages that use the ACCELERATED_COMPOSITING codepath
- Flakiness with the title bar show/hide due to switch to non-embedded
  title bar

Using RealViewSwitcher.java widget from CAF with the following changes:
- Added onScrollStartedListener to notify when a View switch starts
- Added orientationChanged callback

Change-Id: I1f602bbee43b6945a4bac5043ebec8e29379dd3d

Conflicts:

	src/com/android/browser/BrowserSettings.java
	src/com/android/browser/UI.java