ICS browser tab settings changes
Allow the user to toggle whether or not browser tabs are restored on browser launch, either after killing the browser or after reboot. Allow the user to determine the maximum number of allowable tabs instead of using a hard-coded integer.

Patch-set 4: Default setting is 20 but 20 is not available in the list of options from strings.xml. Set the default to 25 in tabs_preferences.xml and Controller.java

Signed-off-by: Gary Danko <gdanko@gmail.com>
Change-Id: If594fdaf55f471f59fd597171f1bfe68ec384f6c

Conflicts:

	res/values/integers.xml